fix: reconcile worker clock display with current consumers
